Group all animations with dom updates
* Better control animation order of execution and re-execution to avoid
  GSAP improperly registering scroll targets
